pyproject.toml
[project]
name = "rag-application"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.4",
    "google-generativeai>=0.8.5",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.0.0",
    "pinecone>=8.0.0",
    "pinecone-client>=6.0.0",
    "pypdf>=6.4.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.1",
    "uvicorn>=0.38.0",
]
